requirements.txt
langchain
langchain-openai
langchain-community
langchain-pinecone
pinecone-client
python-dotenv
yt-dlp
openai-whisper
streamlit
langsmith
tiktoken
fastapi
uvicorn
python-multipart
langchainhub
pypdf
pymupdf
rapidocr-onnxruntime
